Extract logic + handle exception jumps
